# Conflicts:
#	web/sys-web/src/main/java/com/navinfo/dataservice/web/sys/controller/SysMsgController.java

+++ dao/dao-glm-data/src/main/java/com/navinfo/dataservice/dao/glm/model/rd/link/RdLink.java
@@ -102,11 +102,11 @@ public class RdLink implements IObj {
 		this.adasMemo = adasMemo;
 	}
 
-	public int getFeeStd() {
+	public double getFeeStd() {
 		return feeStd;
 	}
 
-	public void setFeeStd(int feeStd) {
+	public void setFeeStd(double feeStd) {
 		this.feeStd = feeStd;
 	}
 
@@ -136,7 +136,7 @@ public class RdLink implements IObj {
 
 	private int truckFlag;
 	
-	private int feeStd;
+	private double feeStd;
 	private int feeFlag;
 	private int systemId;
 	private int originLinkPid;
